requirements.txt
Flask
pydantic
pydantic-settings
imap-tools
python-dotenv 
redis
fastapi
uvicorn
openai
httpx
fastmcp